Moved views into a seperate folder (Just to organize things together)
