+++ micro/cogs/events.py
@@ -5,9 +5,9 @@ from discord.ext import commands
 from classes import (
     Database, 
     GitHub, 
-    CEmbed, 
-    ModerationView
+    CEmbed
 )
+from views import ModerationView
 
 from static.misc import get_extra_channel, get_current_guild
 import static.constants as constants
