+++ micro/views/__init__.py
@@ -1,0 +1,6 @@
+from .issue_list import IssueListView
+from .issue_modal import IssueModal
+from .role import RoleView
+from .resources import ResourcesView
+from .moderation import ModerationView
+from .choose_label import ChooseLabelView
